LayoutTests/fast/hidpi/image-srcset-invalid-inputs-except-one.html

<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
    }

    addEventListener("load", function() {
        shouldBeTrue('document.getElementById("foo").clientWidth==200');
    }, false);
</script>
</head>

<body id="body">
    <div>This test passes if the img tag below is a green square regardless of the scale factor. It ensures that invalid inputs are
    ignored and well-formed images are chosen regardless of their qualifiers, once they are the only candidate left</div>
    <img id="foo" src="" srcset="1x,,  ,   x    , resources/green-200-px-square.png 2x, 2">
</body>
</html>
