safari-600.7-branch/LayoutTests/fast/hidpi/image-srcset-invalid-inputs-except-one-expected.txt

PASS document.getElementById("foo").clientWidth==200 is true
This test passes if the img tag below is a green square regardless of the scale factor. It ensures that invalid inputs are ignored and well-formed images are chosen regardless of their qualifiers, once they are the only candidate left
